backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.28.1
anyio>=4.8.0 